+++ backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/backend/konan/ir/IrUtils.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.kotlin.backend.konan
+package org.jetbrains.kotlin.backend.konan.ir
 
 import org.jetbrains.kotlin.descriptors.*
 import org.jetbrains.kotlin.ir.*
